refactor: split workspace manager and add registry fallbacks
- Break manager.go into focused files (manager, scanner, indexer, migration) and keep the core logic lean

- Keep a persistent workspace registry and make every tool resolve workspaces from it

- index_workspace is now the only tool that walks the filesystem and it has 3 cascading strategies (startup migration, registry fallback, explicit prompt)

- Remove the previous ad-hoc path helpers and simplify the code path for workspace detection
