+++ internal/tools/utils.go
@@ -3,14 +3,12 @@ package tools
 import (
 	"encoding/json"
 	"fmt"
-	"log"
 	"os"
 	"path/filepath"
 	"strings"
 
 	"github.com/doITmagic/rag-code-mcp/internal/codetypes"
 	"github.com/doITmagic/rag-code-mcp/internal/memory"
-	"github.com/doITmagic/rag-code-mcp/internal/workspace"
 )
 
 // readFileLines reads specific lines from a file
@@ -118,51 +116,4 @@ func inferLanguageFromPath(filePath string) string {
 	default:
 		return ""
 	}
-}
-
-// extractFilePathFromParams extracts file path from common parameter names
-func extractFilePathFromParams(params map[string]interface{}) string {
-	pathParams := []string{
-		"file_path",
-		"filePath",
-		"path",
-		"file",
-		"source_file",
-		"target_file",
-	}
-
-	for _, param := range pathParams {
-		if value, ok := params[param]; ok {
-			if path, ok := value.(string); ok && path != "" {
-				return path
-			}
-		}
-	}
-
-	return ""
-}
-
-// resolveFilePathWithFallback extracts file_path from params, falling back to
-// the single active workspace root when only one workspace is indexed.
-// If file_path is resolved via fallback, it is injected into params so that
-// downstream DetectWorkspace calls work transparently.
-// Returns empty string (no error) when no fallback is available, allowing
-// DetectWorkspace/DetectFromParams to use its own CWD fallback.
-func resolveFilePathWithFallback(params map[string]interface{}, wm *workspace.Manager, toolName string) string {
-	filePath := extractFilePathFromParams(params)
-	if filePath != "" {
-		return filePath
-	}
-
-	// Fallback: if exactly one workspace is cached, use its root
-	if wm != nil {
-		if root := wm.GetSingleWorkspaceRoot(); root != "" {
-			log.Printf("[%s] file_path not provided, using single workspace fallback: %s", toolName, root)
-			params["file_path"] = root
-			return root
-		}
-	}
-
-	log.Printf("[%s] file_path not provided and no single workspace fallback available, deferring to DetectWorkspace", toolName)
-	return ""
 }
