+++ internal/tools/workspace_helpers.go
@@ -22,19 +22,10 @@ func CheckCollectionStatus(ctx context.Context, mem memory.LongTermMemory, colle
 		}
 
 		if !exists {
-			// Collection doesn't exist - tell AI to index first
 			return fmt.Sprintf("❌ Workspace '%s' is not indexed yet.\n\n"+
-				"To enable this operation, please call the 'index_workspace' tool first with:\n"+
-				"{\n"+
-				"  \"file_path\": \"%s\"\n"+
-				"}\n\n"+
-				"Details:\n"+
-				"- Workspace: %s\n"+
-				"- Collection: %s (not created yet)\n",
-				workspacePath,
-				workspacePath,
-				workspacePath,
-				collectionName), nil
+				"Please call 'index_workspace' first with {\"file_path\": \"%s\"}.\n"+
+				"Collection: %s (not created yet)",
+				workspacePath, workspacePath, collectionName), nil
 		}
 	}
 
@@ -45,19 +36,10 @@ func CheckCollectionStatus(ctx context.Context, mem memory.LongTermMemory, colle
 // Returns an error message if no results found, nil otherwise.
 func CheckSearchResults(resultCount int, collectionName, workspacePath string) (string, error) {
 	if resultCount == 0 {
-		// Collection exists but is empty - tell AI to index
 		return fmt.Sprintf("❌ Workspace '%s' appears to be empty or not fully indexed.\n\n"+
-			"To enable this operation, please call the 'index_workspace' tool with:\n"+
-			"{\n"+
-			"  \"file_path\": \"%s\"\n"+
-			"}\n\n"+
-			"Details:\n"+
-			"- Workspace: %s\n"+
-			"- Collection: %s (exists but may be empty)\n",
-			workspacePath,
-			workspacePath,
-			workspacePath,
-			collectionName), nil
+			"Please call 'index_workspace' with {\"file_path\": \"%s\"}.\n"+
+			"Collection: %s (exists but may be empty)",
+			workspacePath, workspacePath, collectionName), nil
 	}
 
 	return "", nil
